+++ client/src/components/PageWrapper/index.jsx
@@ -1,22 +1,22 @@
 import React, { useContext } from "react";
 import { UserContext } from "../../context/UserContext";
-import { NavLink, Outlet } from "react-router-dom";
+import { Outlet } from "react-router-dom";
 import Header from "../Header";
 import Navigation from "../Navigation";
 import Message from "../Message";
+import FancyMenu from "../FancyMenu";
 import Footer from "../Footer";
-import { FancyMenu } from "../../pages";
 
 const PageWrapper = () => {
-  const { user } = useContext(UserContext);
+  const { user, login, register, logout } = useContext(UserContext);
   return (
     <>
       <Header />
       <Navigation />
       <Message />
       <Outlet />
-      <Footer />
       {user && <FancyMenu />}
+      <Footer />
     </>
   );
 };
