--- client/src/components/PageWrapper/refactor.jsx
@@ -1,24 +1,0 @@
-import React, { useContext } from "react";
-import { UserContext } from "../../context/UserContext";
-import { Outlet } from "react-router-dom";
-import Header from "../Header";
-import Navigation from "../Navigation";
-import Message from "../Message";
-import FancyMenu from "../FancyMenu";
-import Footer from "../Footer";
-
-const PageWrapper = () => {
-  const { user, login, register, logout } = useContext(UserContext);
-  return (
-    <>
-      <Header />
-      <Navigation />
-      <Message />
-      <Outlet />
-      {user && <FancyMenu />}
-      <Footer />
-    </>
-  );
-};
-
-export default PageWrapper;
