+++ client/src/components/index.jsx
@@ -1,13 +1,16 @@
-export { default as Footer } from "./Footer";
+export { default as PageWrapper } from "./PageWrapper";
 export { default as Header } from "./Header";
 export { default as Message } from "./Message";
 export { default as Navigation } from "./Navigation";
-export { default as PageWrapper } from "./PageWrapper";
-export { default as GitCommitCard } from "./GitCard";
-export { default as StackCard } from "./StackCard";
-export { default as FloatingMenu } from "./FloatingMenu";
-export { default as Button } from "./Button";
+export { default as FancyMeny } from "./FancyMenu";
+export { default as Footer } from "./Footer";
+export { GitCard } from "./GitCard";
+export { StackCard } from "./StackCard";
+export { FloatingMenu } from "./FloatingMenu";
 export { default as CustomLink } from "./CustomLink";
-export { default as Input } from "./Input";
-export { default as Textarea } from "./Textarea";
+export { default as Form_Button } from "./Form_Button";
+export { default as Form_Input } from "./Form_Input";
+export { default as Form_Textarea } from "./Form_Textarea";
+export { default as Form_DropDown } from "./Form_DropDown";
+export { default as Form_Label } from "./Form_Label";
 export { default as Notification } from "./Notification"
